refactored json reader
