+++ app/src/main/java/SmartModerationDesktopApp/Utilities/JsonModerationCardParser.java
@@ -2,7 +2,6 @@ package SmartModerationDesktopApp.Utilities;
 
 import SmartModerationDesktopApp.ModerationCards.ModerationCard;
 import SmartModerationDesktopApp.ModerationCards.ModerationCardData;
-import SmartModerationDesktopApp.Server.LoginInformation;
 import java.awt.Point;
 import java.io.File;
 import java.io.FileReader;
@@ -14,7 +13,7 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
 
-public class JsonReader {
+public class JsonModerationCardParser {
 
     public ArrayList<ModerationCard> parseModerationCardJson(String inputJson) {
         ArrayList<ModerationCard> moderationCardList = new ArrayList();    
@@ -78,12 +77,5 @@ public class JsonReader {
         } catch (ParseException | IOException ex) {
         }
         return cachedModerationCardPositions;
-    }
-
-    public LoginInformation readLoginInformationJson(String loginInformationJson) throws ParseException {
-        JSONParser jsonParser = new JSONParser();
-        JSONObject jsonObject = (JSONObject) jsonParser.parse(loginInformationJson);
-        long port = (long) jsonObject.get("port");
-        return new LoginInformation((long) jsonObject.get("meetingId"), (String) jsonObject.get("ipAddress"), Integer.toString((int) port));
     }
 }
